+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 	
 ### Changes
     
+	- Now you can [un]instrument single vars from the browser, even if they where defined at the repl (Clojure only)
+	- Improved Single var [un]instrumentation from the browser (Clojure and ClojureScript)
+	
 ### Bugs fixed
         
 	- Fix show-error on the clojure local path